requirement.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.3
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-core==2.14.6
pytest==8.0.0
python-dotenv==1.0.1
redis==5.0.1
SQLAlchemy==2.0.25
tomli==2.0.1
typing-extensions==4.9.0
yarl==1.9.4
